Fixed mode bug
The mode variable that is set in line 34 is overridden in the for loop (line 87).
Because of this, mode is always max. For example, in line 98 the suffix will end with max even if we set in line 34 the mode to be min.
To solve this, I used a different variable name for the for loop (I replaced "mode" with "mode_option").